moved discord objects to models/
