+++ asuka/builders/base_objs.py
@@ -23,7 +23,6 @@
 from __future__ import annotations
 
 import datetime
-import typing
 
 
 class DiscordObject:
@@ -45,44 +44,3 @@ class DiscordObject:
         self,
     ) -> datetime.datetime:
         return self._created_at
-
-
-@typing.final
-class Intents:
-    NONE = 0
-    GUILDS = 1 << 0
-    GUILD_MEMBERS = 1 << 1
-    GUILD_BANS = 1 << 2
-    GUILD_EMOJIS = 1 << 3
-    GUILD_INTEGRATIONS = 1 << 4
-    GUILD_WEBHOOKS = 1 << 5
-    GUILD_INVITES = 1 << 6
-    GUILD_VOICE_STATES = 1 << 7
-    GUILD_PRESENCES = 1 << 8
-    GUILD_MESSAGES = 1 << 9
-    GUILD_MESSAGE_REACTIONS = 1 << 10
-    GUILD_MESSAGE_TYPING = 1 << 11
-    DM_MESSAGES = 1 << 12
-    DM_MESSAGE_REACTIONS = 1 << 13
-    DM_MESSAGE_TYPING = 1 << 14
-    MESSAGE_CONTENT = 1 << 15
-    GUILD_SCHEDULED_EVENTS = 1 << 16
-
-    def __init__(
-        self,
-        *args: int,
-    ) -> None:
-        self._value = 0
-        # self.enabled: typing.MutableSet = set()
-        for value in args:
-            self._value |= value
-
-    @property
-    def value(self) -> int:
-        return self._value
-
-    @classmethod
-    def from_value(cls, value: int) -> "Intents":
-        intents = Intents()
-        intents._value = value
-        return intents
