+++ pyproject.toml
@@ -14,6 +14,10 @@ mypy = "^0.971"
 isort = "^5.10.1"
 black = "^22.6.0"
 
+[tool.black]
+max-line-length = 120
+
+
 [build-system]
 requires = ["poetry-core>=1.0.0"]
 build-backend = "poetry.core.masonry.api"
